+++ lib/poolex.ex
@@ -52,9 +52,10 @@ defmodule Poolex do
   """
 
   @typedoc """
-  Any atom naming your pool, e.g. `:my_pool`.
+  Any valid GenServer's name. It may be an atom like `:some_pool` or a tuple {:via, Registry, {MyApp.Registry, "pool"}
+  if you want to use Registry.
   """
-  @type pool_id() :: atom()
+  @type pool_id() :: GenServer.name()
   @typedoc """
   #{@poolex_options_table}
   """
@@ -138,7 +139,7 @@ defmodule Poolex do
       children = [
         Poolex.child_spec(pool_id: :worker_pool_1, worker_module: SomeWorker, workers_count: 5),
         # or in another way
-        {Poolex, [pool_id: :worker_pool_2, worker_module: SomeOtherWorker, workers_count: 5]}
+        {Poolex, pool_id: :worker_pool_2, worker_module: SomeOtherWorker, workers_count: 5}
       ]
 
       Supervisor.start_link(children, strategy: :one_for_one)
@@ -302,13 +303,13 @@ defmodule Poolex do
     waiting_callers_impl =
       Keyword.get(opts, :waiting_callers_impl, Poolex.Callers.Impl.ErlangQueue)
 
-    {:ok, monitor_id} = Monitoring.init(pool_id)
+    {:ok, monitor_pid} = Monitoring.init()
     {:ok, supervisor} = Poolex.Private.Supervisor.start_link()
 
     state =
       %State{
         max_overflow: max_overflow,
-        monitor_id: monitor_id,
+        monitor_pid: monitor_pid,
         pool_id: pool_id,
         supervisor: supervisor,
         worker_args: worker_args,
@@ -347,7 +348,7 @@ defmodule Poolex do
   defp start_workers(workers_count, state) do
     Enum.map(1..workers_count, fn _ ->
       {:ok, worker_pid} = start_worker(state)
-      Monitoring.add(state.monitor_id, worker_pid, :worker)
+      Monitoring.add(state.monitor_pid, worker_pid, :worker)
 
       worker_pid
     end)
@@ -373,13 +374,13 @@ defmodule Poolex do
       if state.overflow < state.max_overflow do
         {:ok, new_worker} = start_worker(state)
 
-        Monitoring.add(state.monitor_id, new_worker, :worker)
+        Monitoring.add(state.monitor_pid, new_worker, :worker)
 
         state = BusyWorkers.add(state, new_worker)
 
         {:reply, {:ok, new_worker}, %State{state | overflow: state.overflow + 1}}
       else
-        Monitoring.add(state.monitor_id, from_pid, :waiting_caller)
+        Monitoring.add(state.monitor_pid, from_pid, :waiting_caller)
 
         state =
           WaitingCallers.add(state, %Poolex.Caller{reference: caller_reference, from: caller})
@@ -464,7 +465,7 @@ defmodule Poolex do
         {:DOWN, monitoring_reference, _process, dead_process_pid, _reason},
         %State{} = state
       ) do
-    case Monitoring.remove(state.monitor_id, monitoring_reference) do
+    case Monitoring.remove(state.monitor_pid, monitoring_reference) do
       :worker ->
         {:noreply, handle_down_worker(state, dead_process_pid)}
 
@@ -512,13 +513,13 @@ defmodule Poolex do
       else
         {:ok, new_worker} = start_worker(state)
 
-        Monitoring.add(state.monitor_id, new_worker, :worker)
+        Monitoring.add(state.monitor_pid, new_worker, :worker)
 
         IdleWorkers.add(state, new_worker)
       end
     else
       {:ok, new_worker} = start_worker(state)
-      Monitoring.add(state.monitor_id, new_worker, :worker)
+      Monitoring.add(state.monitor_pid, new_worker, :worker)
 
       state
       |> BusyWorkers.add(new_worker)
@@ -534,7 +535,7 @@ defmodule Poolex do
   @impl GenServer
   def terminate(reason, %State{} = state) do
     DynamicSupervisor.stop(state.supervisor, reason)
-    Monitoring.stop(state.monitor_id)
+    Monitoring.stop(state.monitor_pid)
 
     :ok
   end