+++ lib/poolex/private/monitoring.ex
@@ -1,40 +1,42 @@
 defmodule Poolex.Private.Monitoring do
   @moduledoc false
-  @type monitor_id() :: atom() | reference()
   @type kind_of_process() :: :worker | :waiting_caller
 
-  @spec init(Poolex.pool_id()) :: {:ok, monitor_id()}
-  @doc false
-  def init(pool_id) do
-    monitor_id = :ets.new(:"#{pool_id}_references", [:set, :named_table, :private])
-
-    {:ok, monitor_id}
+  @spec init() :: {:ok, pid()}
+  @doc """
+  Create new monitoring references storage.
+  """
+  def init do
+    Agent.start_link(fn -> %{} end)
   end
 
-  @spec stop(monitor_id()) :: :ok
-  @doc false
-  def stop(monitor_id) do
-    :ets.delete(monitor_id)
-
-    :ok
+  @spec stop(pid()) :: :ok
+  @doc """
+  Delete storage.
+  """
+  def stop(pid) do
+    Agent.stop(pid)
   end
 
-  @spec add(monitor_id(), pid(), kind_of_process()) :: :ok
-  @doc false
-  def add(monitor_id, process_pid, kind_of_process) do
+  @spec add(monitor_pid :: pid(), worker_pid :: pid(), kind_of_process()) :: :ok
+  @doc """
+  Start monitoring given worker or caller process.
+  """
+  def add(monitor_pid, process_pid, kind_of_process) do
     reference = Process.monitor(process_pid)
-    :ets.insert_new(monitor_id, {reference, process_pid, kind_of_process})
 
-    :ok
+    Agent.update(monitor_pid, fn state -> Map.put(state, reference, kind_of_process) end)
   end
 
-  @spec remove(monitor_id(), reference()) :: kind_of_process()
-  @doc false
-  def remove(monitor_id, monitoring_reference) do
+  @spec remove(monitor_pid :: pid(), reference()) :: kind_of_process()
+  @doc """
+  Stop monitoring given worker or caller process and return kind of it.
+  """
+  def remove(monitor_pid, monitoring_reference) do
     true = Process.demonitor(monitoring_reference)
-    [{_reference, _pid, kind_of_process}] = :ets.lookup(monitor_id, monitoring_reference)
-    true = :ets.delete(monitor_id, monitoring_reference)
 
-    kind_of_process
+    Agent.get_and_update(monitor_pid, fn state ->
+      {Map.get(state, monitoring_reference), Map.delete(state, monitoring_reference)}
+    end)
   end
 end